Escape fro statics in IportOptionsFactory
